pyproject.toml
[tool.poetry]
name = "connectapi"
version = "0.1.0"
description = "Api for konnect"
authors = ["monzero <briankeny15@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.0"
pillow = "^10.1.0"
dj-database-url = "^2.1.0"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
